Present with unknown name when the name string is empty.
I made a previous commit which used to replace name with unknown string but it was not useful as it functioned only in a specific place. I wanted to do for everything.
So this will do that. This commit then overrides the previous commit. We have to remove that commit. It is an extra useless code.