Sample Data Removed. First member default admin.
Earlier I had my family as the default family in the familytree. For any other member you will have to remove all the members and then do any thing else. Now we have
removed that thing. FT now asks the users about the admin and the first member to be added in the family tree.