Solving the null Family Id
We were not passing any family for the admin.
We are aware about the family id of the admin, as it is just being created
before asking the details about the admin member. We have added family id
field in the register form. So if we are aware about the family id of the
member then we add with the other information in the database.`